TestMultipleChainHorcrux, remove rest of old framework
